requirements.txt
fastapi
uvicorn
python-dotenv
openai
twilio
supabase
python-multipart
httpx